Prepare linux3.10 for H5 legacy branch

+++ include/media/davinci/dm644x_ccdc.h
@@ -114,7 +114,7 @@ struct ccdc_fault_pixel {
 	/* Number of fault pixel */
 	unsigned short fp_num;
 	/* Address of fault pixel table */
-	unsigned long fpc_table_addr;
+	unsigned int fpc_table_addr;
 };
 
 /* Structure for CCDC configuration parameters for raw capture mode passed
